🗂 Moved all User queries into the same Resolver
