+++ backend/src/index.ts
@@ -7,15 +7,13 @@ import { buildSchema } from "type-graphql";
 
 import { PORT } from "../config";
 import { BookResolver } from "./modules/book/BookResolver";
-import { LoginResolver } from "./modules/user/LoginResolver";
-import { RegisterResolver } from "./modules/user/RegisterResolver";
 import { UserResolver } from "./modules/user/UserResolver";
 
 import { getUserFromToken } from "./utils";
 
 async function main() {
   await createConnection();
-  const schema = await buildSchema({ resolvers: [BookResolver, LoginResolver, RegisterResolver, UserResolver] });
+  const schema = await buildSchema({ resolvers: [BookResolver, UserResolver] });
   const server = new ApolloServer({
     schema,
     context: ({ req }) => {
