--- backend/src/modules/user/RegisterResolver.ts
@@ -1,25 +1,0 @@
-import { Resolver, Query, Mutation, Arg } from "type-graphql";
-import bcrypt from "bcryptjs";
-
-import { User } from "../../models/User";
-import { RegisterInput } from "./RegisterInput";
-
-@Resolver()
-export class RegisterResolver {
-  @Mutation(() => User)
-  async register(
-    @Arg("data")
-    { email, firstName, lastName, password }: RegisterInput,
-  ): Promise<User> {
-    const hashedPassword = await bcrypt.hash(password, 12);
-
-    const user = await User.create({
-      firstName,
-      lastName,
-      email,
-      password: hashedPassword,
-    }).save();
-
-    return user;
-  }
-}
